requirements.txt
amqp==5.0.2
arabic-reshaper==2.1.0
asgiref==3.2.10
billiard==3.6.3.0
celery==5.0.2
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1.2
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-phonenumber-field==5.0.0
django-timezone-field==4.0
djangorestframework==3.12.1
future==0.18.2
html5lib==1.1
icalendar==4.0.7
importlib-metadata==2.0.0
kombu==5.0.2
phonenumberslite==8.12.11
Pillow==8.0.1
prompt-toolkit==3.0.8
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
reportlab==3.5.53
schedule==0.6.0
six==1.15.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
xhtml2pdf==0.2.5
zipp==3.4.0
